improve: optimize fee calculator
