+++ src/UBAFeeCalculator/index.ts
@@ -1,3 +1,3 @@
 export { default as UBAFeeConfig } from "./UBAFeeConfig";
-export { default as UBAFeeCalculator } from "./UBAFeeCalculator";
+export { default as UBAFeeSpokeCalculator } from "./UBAFeeSpokeCalculator";
 export * as UBAFeeUtility from "./UBAFeeUtility";
